Created preload connection firebase
